fix paging viewer.

+++ modules/file-viewers/src/main/java/org/iplantc/de/fileViewers/client/FileViewer.java
@@ -78,11 +78,11 @@ public interface FileViewer extends IsWidget, IsMaskable, HasHandlers, FileSaved
 
         boolean isDirty();
 
-        void loadPathListData(Integer pageNumber, Integer pageSize, String separator);
+        void loadPathListData(Integer pageNumber, Long pageSize, String separator);
 
-        void loadStructuredData(Integer pageNumber, Integer pageSize, String separator);
+        void loadStructuredData(Integer pageNumber, Long pageSize, String separator);
 
-        void loadTextData(Integer pageNumber, Integer pageSize);
+        void loadTextData(Integer pageNumber, Long pageSize);
 
         void newFileGo(HasOneWidget container,
                        String title,
